requirements.txt
moderngl
pygame
numpy
pyrr
Pillow
texture2ddecoder